Fixed multiprocessing when not using fork
On Windows, forking isn't supported and the spawn method doesn't let the spawned processes inherit state like forking does. That meant that all results to stdout, while multiprocessing, was never redelegated back to the calling, main process.

The SourceVisitor was changed to maintain its output text lines to be extractable with the minimum versions results.

Bonus: Output is now easily tied to the file it originated from. The fork-friendly version just output so the data was spliced across other threads and only when using a single process could one be certain from which file the output corresponded.

Fixes #5.