add --exclude-regex and --no-make-paths-absolute to exclude specific file paths (#115)
* add --exclude-glob argument to exclude file paths from detection

* add --no-make-paths-absolute to allow relative globs

NOTE: tests are broken here because [!/] doesn't work the way I thought it would. The next commit
will change the glob option to use a regex.

* change --exclude-glob to --exclude-regex

- tests pass now
- interface more familiar for windows users

* match using re.search and use anchors

* memoize a "master regex" instead of looping through regex patterns

Add pylint overrides for tests which access the private memoized master regex. I think these tests
are necessary if we think memoizing the regex is a useful optimization, but this may well be
premature optimization.

* Revert "memoize a "master regex" instead of looping through regex patterns"

This reverts commit 27302a3c64802fa72bf9e05bb082f88dfbd9dd20.

* respond to review comments!

* add tests for config parsing

- also add 'yes'/'no' test cases for other boolean config flags

* compare exclusion_regex pattern strings

* fix os path sep for regex on windows