Add unit tests for compare_requirements()
The tests are divided into three categories: one for the case where the
code is Python 2 and 3 compatible, and one each for the cases where the
code is exclusively Python 2 or 3 compatible.

We cover the basic cases of exercising exact and non-exact target
versions. We also cover the issues that come about when the targets can
be missing, or incorrect.

Signed-off-by: Stephen Brennan <stephen@brennan.io>